Вынесет схожую структуру форм в отдельный класс.
